% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriculture_forestry_and_fishery.R
\docType{data}
\name{agriculture_1}
\alias{agriculture_1}
\title{Number of farms by province}
\format{An object of class \code{data.frame} with 1024 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ number_of_farms } A
column of class : integer. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
agriculture_1
}
\description{
Contains the data expressed : 2000-2015, by year and by province.
}
\keyword{datasets}
