% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriculture_forestry_and_fishery.R
\docType{data}
\name{agriculture_2}
\alias{agriculture_2}
\title{Number of farms by kinds of manufacturing sector and by province}
\format{An object of class \code{data.frame} with 315 rows and 7 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ cultivation_farm } A
column of class : integer. \item \code{ fishing_farm } A column of class :
integer. \item \code{ livestock_farm } A column of class : integer. \item
\code{ others } A column of class : integer. \item \code{ total } A column
of class : integer. }}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
agriculture_2
}
\description{
Contains the data expressed : 2011-2015, by year and by province.
}
\keyword{datasets}
