% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriculture_forestry_and_fishery.R
\docType{data}
\name{agriculture_21}
\alias{agriculture_21}
\title{Production of autumn paddy by province}
\format{An object of class \code{data.frame} with 672 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
production_of_autumn_paddy } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam
(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
agriculture_21
}
\description{
Contains the data expressed : 1995-2015, by year and by province.
}
\keyword{datasets}
