% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriculture_forestry_and_fishery.R
\docType{data}
\name{agriculture_35}
\alias{agriculture_35}
\title{Production of peanut by province}
\format{An object of class \code{data.frame} with 504 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ production_of_peanut
} A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
agriculture_35
}
\description{
Contains the data expressed : 1995-2015, by year and by province.
}
\keyword{datasets}
