% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{content}
\alias{content}
\title{Content Table}
\format{A data frame of 316 rows and 8 variables:
\itemize{
   \item \code{category}: global theme of the data frame
   \item \code{subcategory}: subcategory of the data frame
   \item \code{data_frame}: name of the data frame as on the website gso
   \item \code{data_name}: short name of the data frame
   \item \code{time_resolution}: unit of time used in the data frame (year,
   month or single time point)
   \item \code{time_range}: time range expressed in year
   \item \code{sp_resoluation}: spatialdefinition of the data frame
   (province, stations, country, rivers ...)
   \item \code{data}: list of data frame
}}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
data(content)
}
\description{
A tibble containing the description of all the data frame contained in the
package and the data frame in one column \code{data}.
}
\details{
The tibble contains, the names of each data frame,the folder corresponding
which is also the category, a sub category for more information, the
time-range and time-resulotion of the data, the spatial resolution of the
data and the data frame corresponding in a long format.
}
\keyword{datasets}
