% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_culture_and_living_standard.R
\docType{data}
\name{culture_2}
\alias{culture_2}
\title{Number of libraries under local management by province}
\format{An object of class \code{data.frame} with 704 rows and 4 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
number_of_books_in_library_thous_copies } A column of class : numeric. \item
\code{ number_of_libraries_unit } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
culture_2
}
\description{
Contains the data expressed : 2004-2015, by year and by province.
}
\keyword{datasets}
