% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_culture_and_living_standard.R
\docType{data}
\name{culture_3}
\alias{culture_3}
\title{Number of editorial offices by province}
\format{An object of class \code{data.frame} with 504 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
number_of_editorial_offices } A column of class : integer. }}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
culture_3
}
\description{
Contains the data expressed : 2005-2015, by year and by province.
}
\keyword{datasets}
