% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_and_employment.R
\docType{data}
\name{demography_1}
\alias{demography_1}
\title{Area, population and population density by province}
\format{An object of class \code{data.frame} with 15 rows and 4 columns.
\itemize{ \item \code{ year } A column of class : factor. \item \code{ total
} A column of class : numeric. \item \code{ urban } A column of class :
numeric. \item \code{ rural } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
demography_1
}
\description{
Contains the data expressed : 2011-2013, by year and by province.
}
\keyword{datasets}
