% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_and_employment.R
\docType{data}
\name{demography_11}
\alias{demography_11}
\title{Child mortality rate by sex and by residence}
\format{An object of class \code{data.frame} with 640 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ total_fertility_rate
} A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
demography_11
}
\description{
Contains the data expressed : 2001-2015, by year and by residence.
}
\keyword{datasets}
