% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_and_employment.R
\docType{data}
\name{demography_14}
\alias{demography_14}
\title{Under five mortality rate by province}
\format{An object of class \code{data.frame} with 315 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
under_five_mortality_rate } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
demography_14
}
\description{
Contains the data expressed : 2010-2015, by year and by province.
}
\keyword{datasets}
