% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_and_employment.R
\docType{data}
\name{demography_2}
\alias{demography_2}
\title{Average population by sex and by residence}
\format{An object of class \code{data.frame} with 78 rows and 7 columns.
\itemize{ \item \code{ items } A column of class : factor. \item \code{ year
} A column of class : factor. \item \code{ total } A column of class :
numeric. \item \code{ male } A column of class : numeric. \item \code{
female } A column of class : numeric. \item \code{ urban } A column of class
: numeric. \item \code{ rural } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam
(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
demography_2
}
\description{
Contains the data expressed : 1990-2015, by year and by residence.
}
\keyword{datasets}
