% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_and_employment.R
\docType{data}
\name{demography_3}
\alias{demography_3}
\title{Average population by province}
\format{An object of class \code{data.frame} with 62 rows and 6 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ agricultural_production_land } A column of class : numeric. \item
\code{ forestry_land } A column of class : numeric. \item \code{
homestead_land } A column of class : numeric. \item \code{
specially_used_land } A column of class : numeric. \item \code{ total_area }
A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
demography_3
}
\description{
Contains the data expressed : 1995-2015, by year and by province.
}
\keyword{datasets}
