% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_and_employment.R
\docType{data}
\name{demography_4}
\alias{demography_4}
\title{Sex ratio by residence}
\format{An object of class \code{data.frame} with 32 rows and 5 columns.
\itemize{ \item \code{ sex_ratio } A column of class : factor. \item \code{
year } A column of class : factor. \item \code{ total } A column of class :
numeric. \item \code{ urban } A column of class : numeric. \item \code{
rural } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
demography_4
}
\description{
Contains the data expressed : 2006-2015, by year and by residence.
}
\keyword{datasets}
