% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_and_employment.R
\docType{data}
\name{demography_5}
\alias{demography_5}
\title{Sex ratio of population by province}
\format{An object of class \code{data.frame} with 63 rows and 5 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ agricultural_production_land } A column of class : numeric. \item
\code{ forestry_land } A column of class : numeric. \item \code{
homestead_land } A column of class : numeric. \item \code{
specially_used_land } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
demography_5
}
\description{
Contains the data expressed : 2005-2015, by year and by province.
}
\keyword{datasets}
