% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_and_employment.R
\docType{data}
\name{demography_8}
\alias{demography_8}
\title{Crude birth rate, crude death rate and natural increase rate by province}
\format{An object of class \code{data.frame} with 640 rows and 5 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ crude_birth_rate } A
column of class : numeric. \item \code{ crude_death_rate } A column of class
: numeric. \item \code{ natural_increase_rate } A column of class : numeric.
}}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
demography_8
}
\description{
Contains the data expressed : 2005-2015, by year and by province.
}
\keyword{datasets}
