% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_and_employment.R
\docType{data}
\name{divorce_1}
\alias{divorce_1}
\title{Number of divorce cases cleared up in 2015 by province and by level}
\format{An object of class \code{data.frame} with 63 rows and 4 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ district_level } A column of class : integer. \item \code{
provincial_level } A column of class : integer. \item \code{ total } A
column of class : integer. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
divorce_1
}
\description{
Contains the data expressed : 2015, by single time point and by province.
}
\keyword{datasets}
