% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{education_15}
\alias{education_15}
\title{Number of schoolgirls of general education as of 30 September by province}
\format{An object of class \code{data.frame} with 768 rows and 6 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ lower_secondary } A
column of class : integer. \item \code{ primary } A column of class :
integer. \item \code{ total } A column of class : integer. \item \code{
upper_secondary } A column of class : integer. }}
\source{
General Statistical Office of Vietnam
(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
education_15
}
\description{
Contains the data expressed : 2004-2015, by year and by province.
}
\keyword{datasets}
