% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_and_employment.R
\docType{data}
\name{education_2}
\alias{education_2}
\title{Percentage of literate population at 15 years of age and above by province}
\format{An object of class \code{data.frame} with 512 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
percentage_of_literate_population } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam
(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
education_2
}
\description{
Contains the data expressed : 2006-2015, by year and by province.
}
\keyword{datasets}
