% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{education_20}
\alias{education_20}
\title{Number of teachers, students in universities and colleges by province}
\format{An object of class \code{data.frame} with 960 rows and 6 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
of_which_public_student } A column of class : integer. \item \code{
of_which_public_teacher } A column of class : integer. \item \code{ student
} A column of class : integer. \item \code{ teacher } A column of class :
integer. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
education_20
}
\description{
Contains the data expressed : 2001-2015, by year and by province.
}
\keyword{datasets}
