% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{education_24}
\alias{education_24}
\title{Number of teachers and students in professional secondary by province}
\format{An object of class \code{data.frame} with 960 rows and 18 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ 1_student } A column
of class : integer. \item \code{ 1_student_public } A column of class :
integer. \item \code{ 1_teacher } A column of class : integer. \item \code{
1_teacher_public } A column of class : integer. \item \code{ 2_3_student } A
column of class : integer. \item \code{ 2_3_student_public } A column of
class : integer. \item \code{ 2_3_teacher } A column of class : integer.
\item \code{ 2_3_teacher_public } A column of class : integer. \item \code{
3_student } A column of class : integer. \item \code{ 3_student_public } A
column of class : integer. \item \code{ 3_teacher } A column of class :
integer. \item \code{ 3_teacher_public } A column of class : integer. \item
\code{ student } A column of class : integer. \item \code{ student_public }
A column of class : integer. \item \code{ teacher } A column of class :
integer. \item \code{ teacher_public } A column of class : integer. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
education_24
}
\description{
Contains the data expressed : 2001-2015, by year and by province.
}
\keyword{datasets}
