% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education.R
\docType{data}
\name{education_4}
\alias{education_4}
\title{Number of schools, classes, teachers and pupils of kindergarten education as
of 30 September by province}
\format{An object of class \code{data.frame} with 896 rows and 6 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ class } A column of
class : integer. \item \code{ pupil_thous_pupil } A column of class :
integer. \item \code{ school } A column of class : integer. \item \code{
teacher_pers } A column of class : integer. }}
\source{
General Statistical Office of Vietnam
(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
education_4
}
\description{
Contains the data expressed : 2002-2015, by year and by province.
}
\keyword{datasets}
