% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_and_employment.R
\docType{data}
\name{employment_16}
\alias{employment_16}
\title{Percentage of trained employed workers at 15 years of age and above by
province}
\format{An object of class \code{data.frame} with 504 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
percentage_of_trained_employed_workers } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
employment_16
}
\description{
Contains the data expressed : 2008-2015, by year and by province.
}
\keyword{datasets}
