% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_df.R
\name{extract_df}
\alias{extract_df}
\title{Extract data frame from content}
\usage{
extract_df(keyword)
}
\arguments{
\item{keyword}{a vector of character string, of value from the column
`data_name` of the content data frame}
}
\value{
a data frame or a list of data frame
}
\description{
Extracts a data frame or a list of data frame from the name `data_name` of
the data frame selected in the content data frame.
}
\examples{
# for one keyword
pop <- extract_df("demography_5")

# for multiple keyword
lst_demo <- extract_df(c("demography_5", "demography_3"))
}
