% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_keyword.R
\name{find_keyword}
\alias{find_keyword}
\title{Finds keyword}
\usage{
find_keyword(keyword)
}
\arguments{
\item{keyword}{a vector of character string}
}
\value{
a data frame of three variables: \code{data_name}, \code{data_frame}
and \code{data}
}
\description{
Search by keyword in the content data frame and more precisely in the name of
the data frame and in the name of the variables and returns a data frame.
}
\examples{
# for one keyword
pov <- find_keyword("poverty")

# for multiple keyword
export_import <- find_keyword(c("exportation", "importation"))
}
