% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriculture_forestry_and_fishery.R
\docType{data}
\name{fishery_12}
\alias{fishery_12}
\title{Production of aquaculture fish by province}
\format{An object of class \code{data.frame} with 1344 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
production_of_aquaculture_fish } A column of class : integer. }}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
fishery_12
}
\description{
Contains the data expressed : 1995-2015, by year and by province.
}
\keyword{datasets}
