% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriculture_forestry_and_fishery.R
\docType{data}
\name{fishery_6}
\alias{fishery_6}
\title{Production of fishery by province}
\format{An object of class \code{data.frame} with 1344 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
production_of_fishery } A column of class : integer. }}
\source{
General Statistical Office of Vietnam
(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
fishery_6
}
\description{
Contains the data expressed : 1995-2015, by year and by province.
}
\keyword{datasets}
