% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriculture_forestry_and_fishery.R
\docType{data}
\name{forestry_1}
\alias{forestry_1}
\title{Area of forest as of 31 December by province}
\format{An object of class \code{data.frame} with 420 rows and 7 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ area_of_forest } A
column of class : numeric. \item \code{ natural_forest } A column of class :
numeric. \item \code{ newplanted_are } A column of class : numeric. \item
\code{ planted_forest } A column of class : numeric. \item \code{
proportion_of_forest_coverage } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam
(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
forestry_1
}
\description{
Contains the data expressed : 2008-2014, by year and by province.
}
\keyword{datasets}
