% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriculture_forestry_and_fishery.R
\docType{data}
\name{forestry_3}
\alias{forestry_3}
\title{Area of concentrated planted forest by province}
\format{An object of class \code{data.frame} with 1302 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
area_of_concentrated_planted_forest } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam
(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
forestry_3
}
\description{
Contains the data expressed : 1995-2015, by year and by province.
}
\keyword{datasets}
