% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agriculture_forestry_and_fishery.R
\docType{data}
\name{forestry_5}
\alias{forestry_5}
\title{Gross output of wood by province}
\format{An object of class \code{data.frame} with 1344 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ gross_output_of_wood
} A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
forestry_5
}
\description{
Contains the data expressed : 1995-2015, by year and by province.
}
\keyword{datasets}
