% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gso.R
\docType{package}
\name{gso}
\alias{gso}
\alias{gso-package}
\title{gso : Data from the General Statistical Office (GSO) of Vietnam}
\description{
This package contains the data from the General Statistical Office of Vietnam
(last update, September 21st, 2018) ordered in one nested data frame named
"content" which contains the data frame and some complementary information
(spatail & temporal definition ...)
}
