% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_culture_and_living_standard.R
\docType{data}
\name{health_11}
\alias{health_11}
\title{Number of medical staff under provincial departments of health by province}
\format{An object of class \code{data.frame} with 896 rows and 6 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ doctor } A column of
class : integer. \item \code{ midwife } A column of class : integer. \item
\code{ nurse } A column of class : integer. \item \code{ physician } A
column of class : integer. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
health_11
}
\description{
Contains the data expressed : 2002-2015, by year and by province.
}
\keyword{datasets}
