% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_culture_and_living_standard.R
\docType{data}
\name{health_13}
\alias{health_13}
\title{Number of people infected with HIV-AIDS and number of AIDS deaths by
province}
\format{An object of class \code{data.frame} with 315 rows and 7 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
aids_patients_alive_accumulation_as_of_dec_31 } A column of class : integer.
\item \code{ aids_patients_new_case_in } A column of class : integer. \item
\code{ hiv_infected_people_alive_accumulation_as_of_dec_31 } A column of
class : integer. \item \code{ hiv_infected_people_new_case_in } A column of
class : integer. \item \code{ number_of_aids_deathsin } A column of class :
integer. }}
\source{
General Statistical Office of Vietnam
(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
health_13
}
\description{
Contains the data expressed : 2011-2015, by year and by province.
}
\keyword{datasets}
