% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_culture_and_living_standard.R
\docType{data}
\name{health_5}
\alias{health_5}
\title{Number of health establishments under provincial departments of health by
province}
\format{An object of class \code{data.frame} with 704 rows and 7 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ hospital } A column
of class : integer. \item \code{ medical_service_unit } A column of class :
integer. \item \code{ regional_polyclinic } A column of class : integer.
\item \code{ sanatorium_and_rehabilitation_hospital } A column of class :
integer. \item \code{ total } A column of class : integer. }}
\source{
General Statistical Office of Vietnam
(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
health_5
}
\description{
Contains the data expressed : 2005-2015, by year and by province.
}
\keyword{datasets}
