% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrative_unit_land_and_climate.R
\docType{data}
\name{land_use_1}
\alias{land_use_1}
\title{Land use (As of 1 January 2014)}
\format{An object of class \code{data.frame} with 32 rows and 4 columns.
\itemize{ \item \code{ type_of_land } A column of class : factor. \item
\code{ total_area } A column of class : numeric. \item \code{
land_was_allocated_for_users } A column of class : numeric. \item \code{
land_was_allocated_for_managers } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
land_use_1
}
\description{
Contains the data expressed : 2014-01-01, by single time point and by
country.
}
\keyword{datasets}
