% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrative_unit_land_and_climate.R
\docType{data}
\name{land_use_2}
\alias{land_use_2}
\title{Land use by province (As of 1 January 2014)}
\format{An object of class \code{data.frame} with 78 rows and 7 columns.
\itemize{ \item \code{ items } A column of class : factor. \item \code{ year
} A column of class : factor. \item \code{ total } A column of class :
numeric. \item \code{ male } A column of class : numeric. \item \code{
female } A column of class : numeric. \item \code{ urban } A column of class
: numeric. \item \code{ rural } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
land_use_2
}
\description{
Contains the data expressed : 2014-01-01, by single time point and by
province.
}
\keyword{datasets}
