% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrative_unit_land_and_climate.R
\docType{data}
\name{land_use_4}
\alias{land_use_4}
\title{Index of land change in 2014 over 2013 by province (As of annual 1st
January)}
\format{An object of class \code{data.frame} with 45 rows and 5 columns.
\itemize{ \item \code{ residence } A column of class : factor. \item \code{
year } A column of class : factor. \item \code{ crude_birth_rate } A column
of class : numeric. \item \code{ crude_death_rate } A column of class :
numeric. \item \code{ natural_increase_rate } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
land_use_4
}
\description{
Contains the data expressed : 2014-01-01, by single time point and by
province.
}
\keyword{datasets}
