% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_culture_and_living_standard.R
\docType{data}
\name{living_standard_16}
\alias{living_standard_16}
\title{Poverty rate by province}
\format{An object of class \code{data.frame} with 448 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ poverty_rate } A
column of class : numeric. }}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
living_standard_16
}
\description{
Contains the data expressed : 2006-2015, by year and by province.
}
\keyword{datasets}
