% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_culture_and_living_standard.R
\docType{data}
\name{living_standard_26}
\alias{living_standard_26}
\title{Percentage of households in 2014 having house by type of house and by
province}
\format{An object of class \code{data.frame} with 63 rows and 6 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ less_temporary_house } A column of class : numeric. \item \code{
permanent_house } A column of class : numeric. \item \code{
semi_permanent_house } A column of class : numeric. \item \code{
simple_permanent_house } A column of class : numeric. \item \code{ total } A
column of class : numeric. }}
\source{
General Statistical Office of Vietnam
(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
living_standard_26
}
\description{
Contains the data expressed : 2014, by year and by province.
}
\keyword{datasets}
