% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_culture_and_living_standard.R
\docType{data}
\name{living_standard_28}
\alias{living_standard_28}
\title{Average dwelling area per capita in 2014 by type of house and by province}
\format{An object of class \code{data.frame} with 63 rows and 6 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ less_temporary_house } A column of class : numeric. \item \code{
permanent_house } A column of class : numeric. \item \code{
semi_permanent_house } A column of class : numeric. \item \code{
simple_permanent_house } A column of class : numeric. \item \code{ total } A
column of class : numeric. }}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
living_standard_28
}
\description{
Contains the data expressed : 2014, by year and by province.
}
\keyword{datasets}
