% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_culture_and_living_standard.R
\docType{data}
\name{living_standard_3}
\alias{living_standard_3}
\title{Monthly average income per capita at current prices by income source and by
province}
\format{An object of class \code{data.frame} with 189 rows and 7 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
agriculture_forestry_fishery } A column of class : integer. \item \code{
non_agriculture_forestry_fishery } A column of class : integer. \item \code{
others } A column of class : integer. \item \code{ salary_wage } A column of
class : integer. \item \code{ total } A column of class : integer. }}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
living_standard_3
}
\description{
Contains the data expressed : 2010-2014, by year and by province.
}
\keyword{datasets}
