% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_culture_and_living_standard.R
\docType{data}
\name{living_standard_33}
\alias{living_standard_33}
\title{Average collected solid waste treated per day in 2015 by province}
\format{An object of class \code{data.frame} with 63 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ total_of_collected_solid_waste } A column of class : integer. \item
\code{
total_of_collected_solid_waste_treated_in_accordance_with_corresponding_defined_national_criteria
} A column of class : integer. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
living_standard_33
}
\description{
Contains the data expressed : 2015, by year and by province.
}
\keyword{datasets}
