% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_culture_and_living_standard.R
\docType{data}
\name{living_standard_5}
\alias{living_standard_5}
\title{Monthly average income per capita at current prices by income quintile and
by province}
\format{An object of class \code{data.frame} with 189 rows and 8 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ general_average } A
column of class : integer. \item \code{ quintile_1 } A column of class :
integer. \item \code{ quintile_2 } A column of class : integer. \item \code{
quintile_3 } A column of class : integer. \item \code{ quintile_4 } A column
of class : integer. \item \code{ quintile_5 } A column of class : integer. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
living_standard_5
}
\description{
Contains the data expressed : 2010-2014, by year and by province.
}
\keyword{datasets}
