% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_culture_and_living_standard.R
\docType{data}
\name{living_standard_7}
\alias{living_standard_7}
\title{Difference between the highest income quintile and the lowest income
quintile on monthly average income per capita at current prices by province}
\format{An object of class \code{data.frame} with 189 rows and 5 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
monthly_average_income_per_capita_thous_dongs_the_highest_income_quintile }
A column of class : numeric. \item \code{
monthly_average_income_per_capita_thous_dongs_the_lowest_income_quintile } A
column of class : numeric. \item \code{
the_highest_income_quintile_compared_with_the_lowest_income_quintile_time }
A column of class : numeric. }}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
living_standard_7
}
\description{
Contains the data expressed : 2010-2014, by year and by province.
}
\keyword{datasets}
