% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_size}
\alias{pop_size}
\title{Vietnamese population sizes}
\format{A data frame with 1,344 rows and 7 variables:
\itemize{
   \item \code{year}: year
   \item \code{province}: province name
   \item \code{total}: total population
   \item \code{male}: male population
   \item \code{female}: female population
   \item \code{urban}: urban population
   \item \code{rural}: rural population
}}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
pop_size
}
\description{
A dataset containing the Vietnamese population sizes by province (64) and
year from 1995 to 2015.
}
\details{
The precision is 100 people.
}
\keyword{datasets}

