% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_price_and_tourism.R
\docType{data}
\name{price_10}
\alias{price_10}
\title{Spatial cost of living index among provinces (Ha Noi = 100)}
\format{An object of class \code{data.frame} with 315 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
spatial_cost_of_living_index } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
price_10
}
\description{
Contains the data expressed : 2010-2015, by year and by province.
}
\keyword{datasets}
