% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFScoring.R
\name{TFScoring}
\alias{TFScoring}
\title{TFScoring - TF Scoring for evaluation}
\usage{
TFScoring(filepath, algorithm, TFlist, TFtrue, sparSeq)
}
\arguments{
\item{filepath}{Directory path to read output network from GSO algorithms.}

\item{algorithm}{Three group sparse optimization algorithms: Soft, Half and Hard.}

\item{TFlist}{list of candidate TFs.}

\item{TFtrue}{Gold standard TFs.}

\item{sparSeq}{Sequence of group sparsity level.}
}
\description{
TFScoring - TF Scoring for evaluation
}
\details{
The score matrix is generated by all networks with different sparsity.
The TFScoring result consider the first 20 TFs with largest score.
}
\examples{
ntf <- 30
ngene <- 300
nsample <- 20
A <- matrix(rnorm(ntf*nsample), nrow = nsample, ncol = ntf)
B <- matrix(rnorm(ngene*nsample), nrow = nsample, ncol = ngene)
InitialX <- matrix(rnorm(ntf*ngene), nrow = ngene, ncol = ntf)
sparSeq <- c(1:10,seq(10,20,2))
algorithm <- 'Hard'
TFlist <- paste0('G',1:ntf)
TFtrue <- paste0('G',1:10)
resGSO <- demo_GSO(A, B, InitialX, algorithm, sparSeq)
res <- TFScoring('Output/',algorithm = algorithm,TFlist = TFlist,TFtrue = TFtrue,sparSeq = sparSeq)
}
\author{
Jing Qin \href{mailto:qinj29@mail.sysu.edu.cn}{qinj29@mail.sysu.edu.cn}

Xinlin Hu \href{mailto:thompson-xinlin.hu@connect.polyu.hk}{thompson-xinlin.hu@connect.polyu.hk}
}
