% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_price_and_tourism.R
\docType{data}
\name{trade_3}
\alias{trade_3}
\title{Retail sales of goods and services at current prices by province}
\format{An object of class \code{data.frame} with 1344 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
retail_sales_of_goods_and_services } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
trade_3
}
\description{
Contains the data expressed : 1995-2015, by year and by province.
}
\keyword{datasets}
