% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_price_and_tourism.R
\docType{data}
\name{trade_6}
\alias{trade_6}
\title{Number of supermarkets as of annual December 31st by province}
\format{An object of class \code{data.frame} with 488 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
number_of_supermarkets } A column of class : integer. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
trade_6
}
\description{
Contains the data expressed : 2005-2015, by year and by province.
}
\keyword{datasets}
