% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_price_and_tourism.R
\docType{data}
\name{trade_7}
\alias{trade_7}
\title{Number of commercial centers as of annual December 31st by province}
\format{An object of class \code{data.frame} with 360 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
number_of_commercial_centers } A column of class : integer. }}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
trade_7
}
\description{
Contains the data expressed : 2008-2015, by year and by province.
}
\keyword{datasets}
