% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/transport_postal_services_and_telecommunications.R
\docType{data}
\name{transport_18}
\alias{transport_18}
\title{Volume of freight by the road by province}
\format{An object of class \code{data.frame} with 960 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ volume_of_freight }
A column of class : numeric. }}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
transport_18
}
\description{
Contains the data expressed : 2000-2014, by year and by province.
}
\keyword{datasets}
