% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/transport_postal_services_and_telecommunications.R
\docType{data}
\name{transport_21}
\alias{transport_21}
\title{Volume of freight traffic by the waterway by province}
\format{An object of class \code{data.frame} with 810 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
volume_of_freight_traffic } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
transport_21
}
\description{
Contains the data expressed : 2000-2014, by year and by province.
}
\keyword{datasets}
