% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/transport_postal_services_and_telecommunications.R
\docType{data}
\name{transport_7}
\alias{transport_7}
\title{Number of passengers traffic by province}
\format{An object of class \code{data.frame} with 960 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
number_of_passengers_traffic } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam
(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
transport_7
}
\description{
Contains the data expressed : 2000-2014, by year and by province.
}
\keyword{datasets}
