% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/transport_postal_services_and_telecommunications.R
\docType{data}
\name{transport_8}
\alias{transport_8}
\title{Number of passengers carried by the road by province}
\format{An object of class \code{data.frame} with 960 rows and 3 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{
number_of_passengers_carried } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam (\url{http://gso.gov.vn/Default_en.aspx?tabid=491}).
}
\usage{
transport_8
}
\description{
Contains the data expressed : 2000-2014, by year and by province.
}
\keyword{datasets}
