% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_and_employment.R
\docType{data}
\name{demography_1}
\alias{demography_1}
\title{Area, population and population density by province}
\format{An object of class \code{data.frame} with 315 rows and 5 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ area_km2 } A column
of class : numeric. \item \code{ average_population_thous_pers } A column of
class : numeric. \item \code{ population_density_person_km2 } A column of
class : numeric. }}
\source{
General Statistical Office of Vietnam
(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
demography_1
}
\description{
Contains the data expressed : 2011-2013, by year and by province.
}
\keyword{datasets}
