% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_and_employment.R
\docType{data}
\name{demography_11}
\alias{demography_11}
\title{Child mortality rate by sex and by residence}
\format{An object of class \code{data.frame} with 30 rows and 7 columns.
\itemize{ \item \code{ child_mortality_rate } A column of class : factor.
\item \code{ year } A column of class : factor. \item \code{ total } A
column of class : numeric. \item \code{ male } A column of class : numeric.
\item \code{ female } A column of class : numeric. \item \code{ urban } A
column of class : numeric. \item \code{ rural } A column of class : numeric.
}}
\source{
General Statistical Office of Vietnam
(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
demography_11
}
\description{
Contains the data expressed : 2001-2015, by year and by residence.
}
\keyword{datasets}
