% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_and_employment.R
\docType{data}
\name{demography_3}
\alias{demography_3}
\title{Average population by province}
\format{An object of class \code{data.frame} with 1344 rows and 7 columns.
\itemize{ \item \code{ province } A column of class : character. \item
\code{ year } A column of class : integer. \item \code{ female } A column of
class : numeric. \item \code{ male } A column of class : numeric. \item
\code{ rural } A column of class : numeric. \item \code{ total } A column of
class : numeric. \item \code{ urban } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
demography_3
}
\description{
Contains the data expressed : 1995-2015, by year and by province.
}
\keyword{datasets}
