% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/administrative_unit_land_and_climate.R
\docType{data}
\name{land_use_3}
\alias{land_use_3}
\title{Structure of used land by province (As of 1 January 2014)}
\format{An object of class \code{data.frame} with 32 rows and 5 columns.
\itemize{ \item \code{ sex_ratio } A column of class : factor. \item \code{
year } A column of class : factor. \item \code{ total } A column of class :
numeric. \item \code{ urban } A column of class : numeric. \item \code{
rural } A column of class : numeric. }}
\source{
General Statistical Office of Vietnam

(\url{http://gso.gov.vn/Default_en.aspx?tabid=491})
}
\usage{
land_use_3
}
\description{
Contains the data expressed : 2014-01-01, by single time point and by
province.
}
\keyword{datasets}
